fit application status and delete user pop up under base modal and base alert
