+++ src/components/bases/BaseModal.tsx
@@ -13,7 +13,7 @@ type BaseModalProps = {
 export default function BaseModal({ open, onClose, children, title }: BaseModalProps) {
   return (
     <Modal open={open} onClose={onClose}>
-      <Box className="flex flex-col justify-between absolute top-1/2 left-1/2 -translate-x-1/2 -translate-y-1/2 bg-white rounded-sm p-6 w-[600px] h-[191px]">
+      <Box className="flex flex-col justify-between absolute top-1/2 left-1/2 -translate-x-1/2 -translate-y-1/2 bg-white rounded-sm p-6 w-[600px] min-h-[191px]">
         {title && <h2 className="text-xl font-semibold mb-4 text-[#123A1E]">{title}</h2>}
         {children}
       </Box>
